+++ src/execution/execute_command.c
@@ -6,7 +6,7 @@
 /*   By: mkakizak <mkakizak@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/28 09:57:08 by minoka            #+#    #+#             */
-/*   Updated: 2024/12/09 16:51:25 by mkakizak         ###   ########.fr       */
+/*   Updated: 2024/12/09 17:48:39 by mkakizak         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -89,7 +89,9 @@ char	**set_command(char *command, char **args)
 
 
 
-int	execute_cmd(t_command *cmd, t_cmnd_tbl *table, int is_child, char *envp[])
+// restoreing file descriptors seems to work
+
+int	execute_cmd(t_command *cmd, t_cmnd_tbl *table, int is_child, char *envp[], t_fd *fd)
 {
 	char	**set_array;
 	char	*path;
@@ -102,12 +104,14 @@ int	execute_cmd(t_command *cmd, t_cmnd_tbl *table, int is_child, char *envp[])
 	}
 
 	if (path == NULL && !is_path(cmd->command))
-	{
+	{	
+		restore_fd(fd);
 		ft_printf("%s: command not found\n", cmd->command);
 		exit(127);
 	}
 	if (is_directory(path))
 	{
+		restore_fd(fd);
 		ft_printf("minishell: %s: is a directory\n", cmd->command);
 		exit(126);
 	}
@@ -115,8 +119,8 @@ int	execute_cmd(t_command *cmd, t_cmnd_tbl *table, int is_child, char *envp[])
 
 	if (execve(path, set_array, envp) == -1)
 	{	
+		restore_fd(fd);
 		ft_printf("minishell: %s: %s\n", cmd->command, strerror(errno));
-		// free_command_list(cmd);
 		exit(127);
 	}
 	return (0);
