+++ src/execution/pipeline.c
@@ -6,101 +6,11 @@
 /*   By: mkakizak <mkakizak@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/28 13:03:02 by minoka            #+#    #+#             */
-/*   Updated: 2024/12/09 17:02:08 by mkakizak         ###   ########.fr       */
+/*   Updated: 2024/12/09 17:49:02 by mkakizak         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <minishell.h>
-
-
-//TODO: i need a way to know which is the last redirection in the list 
-	// like if i have echo hello > file1 >> file2 > file3
-	//then hello should be written in the file3
-
-int output_redirect(t_command *cmd)
-{
-	int fd;
-	int flags;
-	int i;
-
-	if(cmd->output_file == NULL || cmd->output_file[0] == NULL)
-		return(0);
-
-
-	i = 0;
-	while(cmd->output_file[i])
-	{
-		//output file is >
-		flags = O_WRONLY | O_CREAT ;
-
-		fd = open(cmd->output_file[i], flags, 0644);
-		if(fd == -1)
-		{	
-			ft_printf("minishell: %s: %s\n", cmd->output_file[i], strerror(errno));
-			exit(EXIT_FAILURE);
-		}
-		dup2(fd, STDOUT_FILENO);
-		close(fd);
-		i++;
-	}
-	return (0);
-}
-
-int append_redirect(t_command *cmd)
-{
-	int fd;
-	int flags;
-	int i;
-
-	if(cmd->append == NULL || cmd->append[0] == NULL)
-		return(0);
-
-	i = 0;
-	while(cmd->append[i])
-	{
-		//append is >> 
-		flags = O_WRONLY | O_CREAT | O_APPEND;
-
-		fd = open(cmd->append[i], flags, 0644);
-		if(fd == -1)
-		{
-			ft_printf("minishell: %s: %s\n", cmd->append[i], strerror(errno));
-			exit(EXIT_FAILURE);
-		}
-		dup2(fd, STDOUT_FILENO);
-		close(fd);
-		i++;
-	}
-	return (0);
-}
-
-
-int input_redirect(t_command *cmd)
-{
-	int fd;
-	int flags;
-	int i;
-
-	if(cmd->input_file == NULL || cmd->input_file[0] == NULL)
-		return(0);
-
-	i = 0;
-	while(cmd->input_file[i])
-	{
-		flags = O_RDONLY;
-
-		fd = open(cmd->input_file[i], flags, 0644);
-		if(fd == -1)
-		{
-			ft_printf("minishell: %s: %s\n", cmd->input_file[i], strerror(errno));
-			exit(EXIT_FAILURE);
-		}
-		dup2(fd, STDIN_FILENO);
-		close(fd);
-		i++;
-	}
-	return (0);
-}
 
 int setup_pipes(int *prev_pipe, t_command *current, t_fd *fd)
 {
@@ -156,53 +66,8 @@ void await_process(pid_t pid, t_cmnd_tbl *table)
 	return ;
 }
 
-// typedef struct s_command
-// {
-// 	char				*command;
-// 	char				**args;
-// 	char				**input_file;
-// 	char				**output_file;
-// 	char				**heredoc_delimiter;
-// 	char				**append;
-// 	int					last_input_heredoc;
-// 	int					last_output_append;
-// 	int					pipe_in;
-// 	int					pipe_out;
-// 	int					is_built_in;
-// 	struct s_command	*next;
-// }						t_command;
-
-
-int redirects(t_command *current, t_fd *fd, int prev_pipe)
-{
-	setup_pipes(&prev_pipe, current, fd);
-	if(current->last_input_heredoc == HEREDOC)
-	{
-		input_redirect(current);
-		heredoc_redirect(current);
-	}
-	else
-	{
-		heredoc_redirect(current);
-		input_redirect(current);
-	}
-	if(current->last_output_append == APPEND)
-	{
-		output_redirect(current);
-		append_redirect(current);
-	}
-	else
-	{
-		output_redirect(current);
-		append_redirect(current);
-	}
-	return (0);
-}
-
-
 int	pipeline(t_cmnd_tbl *table, char *envp[])
 {
-	//maybe i can add some of these to the table struct
 	t_fd 		fd;
 	pid_t 		pid;
 	t_command	*current;
@@ -217,11 +82,8 @@ int	pipeline(t_cmnd_tbl *table, char *envp[])
 	pid = -1;
 	init_fd(&fd);
 
-
-
 	while(current)
 	{
-
 		if(current->next)
 			init_pipe(&fd);
 
@@ -235,24 +97,21 @@ int	pipeline(t_cmnd_tbl *table, char *envp[])
 			is_child = TRUE;
 			redirects(current, &fd, prev_pipe);
 			if(current->is_built_in)
-				built_in_cmds(current, table, is_child);
+				built_in_cmds(current, table, is_child, &fd);
 			else
-				execute_cmd(current, table, is_child, envp);
+				execute_cmd(current, table, is_child, envp, &fd);
 		}
 		else if (current->is_built_in && !has_pipe(table->head))
 		{
 
-			built_in_cmds(current, table, is_child);
+			built_in_cmds(current, table, is_child, &fd);
 		}
-
+		
 		clean_pipes(&prev_pipe, current, &fd);
 		current = current->next;
 	}
 
 	await_process(pid, table);
-	// puts("end of piepline\n");
-	// free_command_list(table->head);
-	// printf("exit status is : %d\n", table->last_exit_status);
 	restore_fd(&fd);
 	return (0);
 }
